break models out to individual files with struct def and supporting functions
