+++ internal/in/get_params.go
@@ -1,0 +1,12 @@
+package in
+
+import (
+	"fmt"
+)
+
+type GetParams struct {
+}
+
+func (p GetParams) ValidateParams() error {
+	return fmt.Errorf("not implemented yet")
+}
